requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0.1
django-crispy-forms==1.13.0
mysqlclient==2.1.0
Pillow==9.0.0
sqlparse==0.4.2
djangorestframework~=3.13.1
